bulk-upload-v3/requirements.txt
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.50.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.7
paramiko==3.4.0
piexif==1.1.3
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyNaCl==1.5.0
requests==2.31.0
rsa==4.9
setuptools==68.2.2
SQLAlchemy==2.0.29
sshtunnel==0.4.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.41.2
